+++ components/shared/container.tsx
@@ -2,10 +2,15 @@ import React, { FC } from "react";
 
 interface Props {
   children: React.ReactNode;
+  className?: string;
 }
 
-const Container: FC<Props> = ({ children }) => {
-  return <div className="p-2 w-full md:max-w-6xl mx-auto">{children}</div>;
+const Container: FC<Props> = ({ children, className = "" }) => {
+  return (
+    <div className={`p-2 w-full md:max-w-6xl mx-auto ${className}`}>
+      {children}
+    </div>
+  );
 };
 
 export default Container;